phase_3_predictive_analytics/requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
joblib>=1.3
kafka-python>=2.0.2
statsmodels>=0.14
pyyaml>=6.0
